Separated testing from the example geometry
