+++ src/t8_geometry/t8_geometry_implementations/t8_geometry_lagrange.cxx
@@ -19,12 +19,17 @@
   along with t8code; if not, write to the Free Software Foundation, Inc.,
   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
+#include <array>
+#include <sstream>
 
 #include <t8_geometry/t8_geometry_base.hxx>
 #include <t8_cmesh/t8_cmesh_types.h>
 #include <t8_geometry/t8_geometry_implementations/t8_geometry_lagrange.hxx>
 #include <t8_geometry/t8_geometry_implementations/t8_geometry_lagrange.h>
 #include <t8_eclass.h>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
+#include <t8_forest/t8_forest_general.h>
+#include <t8_forest/t8_forest_io.h>
 
 t8_geometry_lagrange::t8_geometry_lagrange (int dim): t8_geometry_with_vertices (dim, "")
 {
@@ -231,6 +236,250 @@ t8_geometry_lagrange::t8_geom_h27_basis (const double *ref_point) const
         32 * eta * xi * zeta * (eta - 1) * (xi - 1) * (zeta - 0.5),
         -64 * eta * xi * zeta * (eta - 1) * (xi - 1) * (zeta - 1) };
   return basis_functions;
+}
+
+t8_forest_t
+LagrangeElement::create_uniform_forest (t8_cmesh_t cmesh, uint level) const
+{
+  t8_forest_t forest;
+  forest = t8_forest_new_uniform (cmesh, t8_scheme_new_default_cxx (), level, 0, sc_MPI_COMM_WORLD);
+  return forest;
+};
+
+LagrangeElement::LagrangeElement (t8_eclass_t eclass, uint degree, std::vector<double> &nodes)
+  : eclass (eclass), degree (degree), nodes (nodes)
+{
+  // TODO: Check if the number of nodes corresponds to the element type and degree.
+  // if (nodes.size () != parametric_nodes.size ())
+  //   SC_ABORTF ("Provide the 3 coordinates of the nodes.\n");
+  /* Create a cmesh with a single element */
+  int dim = t8_eclass_to_dimension[eclass];
+  t8_geometry_c *geometry = new t8_geometry_lagrange (dim);  // TODO: do it with a smart pointer or using RAII
+  t8_cmesh_init (&cmesh);
+  t8_cmesh_set_attribute (cmesh, 0, t8_get_package_id (), T8_CMESH_LAGRANGE_POLY_DEGREE, &degree, sizeof (int), 1);
+  t8_cmesh_register_geometry (cmesh, geometry);
+  t8_cmesh_set_tree_class (cmesh, 0, eclass);
+  t8_cmesh_set_tree_vertices (cmesh, 0, nodes.data (), nodes.size ());
+  t8_cmesh_commit (cmesh, sc_MPI_COMM_WORLD);
+}
+
+const uint LagrangeElement::lagrange_nodes[T8_ECLASS_COUNT][2];
+
+t8_eclass_t
+LagrangeElement::getType () const
+{
+  return eclass;
+}
+
+std::vector<std::vector<uint>>
+LagrangeElement::getFaceNodes () const
+{
+  std::ostringstream invalid_degree;
+  invalid_degree << "Invalid degree " << degree << ".\n";
+  std::vector<std::vector<uint>> face_nodes;
+  switch (eclass) {
+  case T8_ECLASS_LINE:
+    if (degree == 1)
+      face_nodes = { { 0 }, { 1 } };
+    else if (degree == 2)
+      face_nodes = { { 0 }, { 1 }, { 2 } };
+    else
+      SC_ABORTF (invalid_degree.str ().c_str ());
+    break;
+  case T8_ECLASS_TRIANGLE:
+    if (degree == 1)
+      face_nodes = { { 1, 2 }, { 2, 0 }, { 0, 1 } };
+    else if (degree == 2)
+      face_nodes = { { 1, 2, 3 }, { 2, 0, 4 }, { 0, 1, 5 } };
+    else
+      SC_ABORTF (invalid_degree.str ().c_str ());
+    break;
+  case T8_ECLASS_QUAD:
+    if (degree == 1) {
+      face_nodes = { { 2, 0 }, { 1, 3 }, { 0, 1 }, { 3, 2 } };
+    }
+    else if (degree == 2)
+      face_nodes = { { 2, 0, 4 }, { 1, 3, 5 }, { 0, 1, 6 }, { 3, 2, 7 } };
+    else
+      SC_ABORTF (invalid_degree.str ().c_str ());
+    break;
+  case T8_ECLASS_HEX:
+    if (degree == 1) {
+      face_nodes = { { 2, 0, 6, 4 }, { 1, 3, 5, 7 }, { 0, 1, 4, 5 }, { 3, 2, 7, 6 }, { 2, 3, 0, 1 }, { 4, 5, 6, 7 } };
+    }
+    else if (degree == 2) {
+      face_nodes
+        = { { 2, 0, 6, 4, 8, 9, 10, 11, 12 },  { 1, 3, 5, 7, 13, 14, 15, 16, 17 }, { 0, 1, 4, 5, 9, 13, 18, 19, 20 },
+            { 3, 2, 7, 6, 14, 8, 21, 22, 23 }, { 2, 3, 0, 1, 10, 15, 21, 18, 24 }, { 4, 5, 6, 7, 11, 16, 19, 22, 25 } };
+    }
+    else
+      SC_ABORTF (invalid_degree.str ().c_str ());
+    break;
+  default:
+    SC_ABORTF ("Invalid element class %d.\n", eclass);
+  }
+  return face_nodes;
+}
+
+std::vector<t8_eclass_t>
+LagrangeElement::faceClasses () const
+{
+  std::vector<t8_eclass_t> face_classes;
+  switch (eclass) {
+  case T8_ECLASS_LINE:
+    face_classes = std::vector<t8_eclass_t> (2, T8_ECLASS_VERTEX);
+    break;
+  case T8_ECLASS_QUAD:
+    face_classes = std::vector<t8_eclass_t> (4, T8_ECLASS_LINE);
+    break;
+  case T8_ECLASS_TRIANGLE:
+    face_classes = std::vector<t8_eclass_t> (3, T8_ECLASS_LINE);
+    break;
+  case T8_ECLASS_HEX:
+    face_classes = std::vector<t8_eclass_t> (6, T8_ECLASS_QUAD);
+    break;
+  case T8_ECLASS_TET:
+    face_classes = std::vector<t8_eclass_t> (4, T8_ECLASS_TRIANGLE);
+    break;
+  case T8_ECLASS_PRISM:
+    face_classes = { T8_ECLASS_QUAD, T8_ECLASS_QUAD, T8_ECLASS_QUAD, T8_ECLASS_TRIANGLE, T8_ECLASS_TRIANGLE };
+    break;
+  case T8_ECLASS_PYRAMID:
+    face_classes = { T8_ECLASS_TRIANGLE, T8_ECLASS_TRIANGLE, T8_ECLASS_TRIANGLE, T8_ECLASS_TRIANGLE, T8_ECLASS_QUAD };
+    break;
+  default:
+    SC_ABORTF ("Invalid element class %d.\n", eclass);
+  }
+  return face_classes;
+}
+
+std::vector<double>
+LagrangeElement::getNodeCoords (uint node) const
+{
+  const double *v = t8_cmesh_get_tree_vertices (cmesh, 0);
+  return std::vector<double> (v + 3 * node, v + 3 * node + 3);
+}
+
+std::vector<std::vector<double>>
+LagrangeElement::getNodeCoords (std::vector<uint> &nodes) const
+{
+  const double *v = t8_cmesh_get_tree_vertices (cmesh, 0);
+  size_t n_node = nodes.size ();
+  std::vector<std::vector<double>> node_coords (n_node);
+  for (size_t i = 0; i < n_node; ++i) {
+    uint i_node = nodes[i];
+    node_coords[i] = std::vector<double> (v + 3 * i_node, v + 3 * i_node + 3);
+  }
+  return node_coords;
+}
+
+std::vector<LagrangeElement>
+LagrangeElement::decompose () const
+{
+  /* Get the node numbers of the faces */
+  std::vector<t8_eclass_t> fc = faceClasses ();
+  std::vector<std::vector<uint>> fn = getFaceNodes ();
+  /* Create a new Lagrange element from each face */
+  std::vector<LagrangeElement> faces;
+  const uint n_face = t8_eclass_num_faces[eclass];
+  faces.reserve (n_face);
+  for (size_t i_face = 0; i_face < n_face; ++i_face) {
+    auto nc = flatten<double> (getNodeCoords (fn[i_face]));
+    faces.emplace_back (fc[i_face], degree, nc);
+  }
+  return faces;
+}
+
+std::array<double, 3>
+LagrangeElement::evaluate (const std::array<double, 3> &ref_point) const
+{
+  std::array<double, 3> mapped;
+  t8_geometry_evaluate (cmesh, 0, ref_point.data (), 1, mapped.data ());
+  return mapped;
+}
+
+std::array<double, 3>
+LagrangeElement::mapOnFace (t8_eclass map_onto, const uint face_id, const std::array<double, 3> &coord) const
+{
+  /* Error messages for input validation */
+  std::ostringstream unsupported_element;
+  unsupported_element << "Mapping from a " << t8_eclass_to_string[eclass] << " element is not supported yet.\n";
+  std::ostringstream unsupported_target_element;
+  unsupported_target_element << "Mapping onto a " << t8_eclass_to_string[map_onto]
+                             << " element is not supported yet.\n";
+  std::ostringstream too_many_faces;
+  too_many_faces << "A " << t8_eclass_to_string[map_onto] << " element has " << t8_eclass_num_faces[map_onto]
+                 << " faces only.\n";
+  std::ostringstream non_matching_elem_classes;
+  non_matching_elem_classes << "Face " << face_id << " of a " << t8_eclass_to_string[map_onto]
+                            << " element is not of type " << t8_eclass_to_string[eclass] << ".\n";
+  if (face_id > t8_eclass_num_faces[map_onto] - 1)
+    SC_ABORT (too_many_faces.str ().c_str ());
+
+  /* Actual mapping, case by case */
+  std::array<double, 3> mapped_coord;
+  double xi = coord[0];
+  double eta = coord[1];
+  double zeta = coord[2];
+  switch (eclass) {
+  case T8_ECLASS_LINE:
+    if (map_onto == T8_ECLASS_TRIANGLE) {
+      if (face_id == 0)
+        mapped_coord = { 1, xi, 0 };
+      else if (face_id == 1)
+        mapped_coord = { 1 - xi, 1 - xi, 0 };
+      else if (face_id == 2)
+        mapped_coord = { xi, 0, 0 };
+    }
+    else if (map_onto == T8_ECLASS_QUAD) {
+      if (face_id == 0)
+        mapped_coord = { 0, 1 - xi, 0 };
+      else if (face_id == 1)
+        mapped_coord = { 1, xi, 0 };
+      else if (face_id == 2)
+        mapped_coord = { xi, 0, 0 };
+      else if (face_id == 3)
+        mapped_coord = { 1 - xi, 1, 0 };
+    }
+    else
+      SC_ABORT (unsupported_target_element.str ().c_str ());
+    break;
+  case T8_ECLASS_QUAD:
+    if (map_onto == T8_ECLASS_HEX) {
+      if (face_id == 0)
+        mapped_coord = { 0, 1 - xi, eta };
+      else if (face_id == 1)
+        mapped_coord = { 1, xi, eta };
+      else if (face_id == 2)
+        mapped_coord = { xi, 0, eta };
+      else if (face_id == 3)
+        mapped_coord = { 1 - xi, 1, eta };
+      else if (face_id == 4)
+        mapped_coord = { xi, 1 - eta, 0 };
+      else if (face_id == 5)
+        mapped_coord = { xi, eta, 1 };
+    }
+    else
+      SC_ABORT (unsupported_target_element.str ().c_str ());
+    break;
+  default:
+    SC_ABORT (unsupported_element.str ().c_str ());
+  }
+  return mapped_coord;
+}
+
+void
+LagrangeElement::write () const
+{
+  /* A cmesh cannot be exported, only a forest.
+       So we create one tree element per coarse mesh element */
+  t8_forest_t forest = create_uniform_forest (cmesh, 0);
+  std::ostringstream filename;
+  filename << "Lagrange" << t8_eclass_to_string[eclass] << "Degree" << degree;
+  t8_forest_write_vtk (forest, filename.str ().c_str ());
+  /* Clean up the dummy forest, making sure that the cmesh is not destroyed */
+  t8_cmesh_ref (cmesh);
+  t8_forest_unref (&forest);
 }
 
 T8_EXTERN_C_BEGIN ();
